refactor: remove the concept of passing booleans to the lint level
