+++ source-code/core/src/lint/context.ts
@@ -1,6 +1,6 @@
 import type { Resource, Message, Pattern } from "../ast/schema.js"
 import type { Pretty } from "../utilities/types.js"
-import type { LintableNode, LintConfigArguments, LintRuleId } from "./rule.js"
+import type { LintableNode, LintRuleId } from "./rule.js"
 
 export type LintLevel = "error" | "warn"
 
@@ -34,25 +34,4 @@ export type Context = {
 	reference?: LintableNode
 	referenceLanguage: string
 	languages: string[]
-}
-
-/**
- * An utility function to parse the arguments passed to a lint rule. It will extract the lint rule and the rule specific settings.
- *
- * @param args the settings object pass
- * @param defaultLevel the fallback lint level that get's used when no lint level is specified
- * @returns the extracted information from the passed arguments
- */
-export const parseLintConfigArguments = <T>(
-	args: LintConfigArguments<T> | undefined,
-	defaultLevel: LintLevel,
-): { level: LintLevel; settings: T | undefined } => {
-	const [parsedLevel, settings] = args || []
-
-	const level = parsedLevel === undefined || parsedLevel === true ? defaultLevel : parsedLevel
-
-	return {
-		level,
-		settings,
-	}
 }
