+++ source-code/core/src/lint/index.ts
@@ -1,5 +1,5 @@
 export { lint } from "./linter.js"
 export { getLintReports, hasLintReports } from "./query.js"
-export { type Context, parseLintConfigArguments } from "./context.js"
+export { type Context } from "./context.js"
 export { createLintRule, type LintRule, type LintRuleInitializer, type LintRuleId } from "./rule.js"
 export { createLintRuleCollection, type RuleCollectionInitializer } from "./ruleCollection.js"
